# Conflicts:
#   tools/benchmark/src/main/java/fr/umlv/rootcause/benchmark/Benchmark.java

+++ core/src/main/java/fr/uge/modules/linking/token/type/TypeTime.java
@@ -25,7 +25,7 @@ public class TypeTime implements TokenType{
     }
 
     @Override
-    public float computeProximity(List<TokenEntity> tokenLeft, List<TokenEntity> tokenRight) {
+    public double computeProximity(List<TokenEntity> tokenLeft, List<TokenEntity> tokenRight) {
         return 0;
     }
